Button Menu Bar and Slideshow
New menu bar for some functions like Screenshot, Persist, Exit, Clean...
New Slideshow with all the files sent (screenshots is the most interesting). In this case we shuld filter only images for taking the images and screenshots. (Watching the code, I saw that the screenshots name start with a prefix but it isnt a pattern to llok for cause any other file could be named like that and if i filter with that prefix could by bad practise... have to think on this)

In a future would create a Drag & Drop FileInput for sending files to the agent.